static-stack: Update for upstream-nixpkgs-musl-1.1.19 and the stack v1.9 branch.
This works for stack commit a2489de02 with Nix 2.0.4 on Ubuntu.

Sn survey, also switch witch from `dontCheck = true` to `doCheck = false` for Cython.
The old way seemed to be working for Python 2.7 as done there, but
for Python 3 as needed for static-stack it seems to have no effect, so I made both
use `doCheck` which works fine.