static-stack: Make it work with stack-1.9.1.
Works on stack commit f9d0042c141.

The changes are necessary because stack-1.9.1 explicitly
depends on Cabal-2.4.0.1 in its stack.yaml which is pulled in
via stack2nix.

Doing this, I also found that the way I patched Cabal so far was not
really optimal, because not only for stack, but in general my patches
for Cabal wouldn't be used if the Cabal version wasn't the default one in
nixpkgs.

Now instead I use patches instead of a Github fork so that they can
be applied to any Cabal version.
Also added is a framework for backwards compatibility in the situation
that upstream Cabal starts merging my patches and they appear in nixpkgs.

I've also checked that there's no regression to the set of all Stackage
executables that can be built (#4) and added the non-working ones explicitly
into the survey file.