WIP: static-stack: Make it work with stack-1.9.1.
stack commit f9d0042c141.

The changes are necessary because stack-1.9.1 explicitly
depends on Cabal-2.4.0.1 in its stack.yaml which is pulled in
via stack2nix.

WIP because it breaks building the version of stack that is
in nixpkgs.

Also the handling of `builtins.isNull super.Cabal` is super dodgy.
Not sure why it is necessary (it isn't for the 1.9 build but
completely breaks non-1.9 builds).