static-stack2nix-builder: Improve it a bit.
I found that the `run-stack2nix-and-static-build-script`
is just not a good idea, with its assumption that it runs in a dir
where some `default.nix` with some `static_package` is present.

It just breaks too easily, and is rather confusing, so we'll let
the consumers of static-stack2nix-builder write something like that
themselves if they want it.