Update to latest nixpkgs unstable.
This switches from my nixpkgs fork back to upstream nixpkgs,
which thanks to community-sponsored CI has been shown to be
green for a while.

Not carrying my changes to `pkgsStatic` breaks support for
using that as a base, but that was only an experiment on
my side anyway. `pkgsMusl` is the better base.